requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0
python-dotenv==1.0.0
slack-sdk==3.26.1
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
requests==2.31.0
pgvector==0.2.4
alembic==1.12.1
pytest==7.4.3
pytest-cov==4.1.0
python-json-logger==2.0.7

